requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
aiohttp==3.9.0
python-dotenv==1.0.0
python-multipart==0.0.9
stripe==7.4.0
qrcode==7.4.2
pillow==11.0.0